Ensure tox gets proper language settings.
If LC_ALL is not passed then the interpreter will assume it is C
(which causes the interpreter to use ASCII). This, in turn, causes
the unicode in various places to trigger a UnicodeDecodeError
exception.

CI workers have that locale set by default.
Some people do not have that locale setting set properly.