Allow display of custom message when no-commit-to-branch fails
This change should make the failures of no-commit-to-branch easier
for people that might encounter it as maintainers can provide
information about allowed branch names. Current behavior just gives
an opaque error that is not very useful for people that do not have
a lot of knowledge about the project they are contributing to.